Refactor code by moving UserPanel to its own file
